requirements.txt
google-cloud-speech
openai
numpy
scikit-learn
pandas
